pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.7.1",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
